+++ backend/src/main/kotlin/org/loculus/backend/service/submission/FileMappingPreconditionValidator.kt
@@ -1,0 +1,58 @@
+package org.loculus.backend.service.submission
+
+import org.loculus.backend.api.FileCategory
+import org.loculus.backend.api.FileCategoryFilesMap
+import org.loculus.backend.api.categories
+import org.loculus.backend.api.getDuplicateFileNames
+import org.loculus.backend.api.Organism
+import org.loculus.backend.api.SubmissionIdFilesMap
+import org.loculus.backend.config.BackendConfig
+import org.loculus.backend.controller.UnprocessableEntityException
+import org.springframework.stereotype.Component
+
+@Component
+class FileMappingPreconditionValidator(private val backendConfig: BackendConfig) {
+    fun validateFilenamesAreUnique(submissionIdFilesMap: SubmissionIdFilesMap?): FileMappingPreconditionValidator {
+        submissionIdFilesMap?.values?.forEach {
+            this.validateFilenamesAreUnique(it)
+        }
+        return this
+    }
+
+    fun validateFilenamesAreUnique(fileCategoriesFilesMap: FileCategoryFilesMap?): FileMappingPreconditionValidator {
+        if (fileCategoriesFilesMap == null) return this
+        fileCategoriesFilesMap.categories().forEach { category: FileCategory ->
+            val duplicateFileNames = fileCategoriesFilesMap.getDuplicateFileNames(category)
+            if (duplicateFileNames.isNotEmpty()) {
+                throw UnprocessableEntityException(
+                    "The files in category $category contain duplicate file names: ${duplicateFileNames.joinToString()}",
+                )
+            }
+        }
+        return this
+    }
+
+    fun validateCategoriesMatchSchema(submissionIdFilesMap: SubmissionIdFilesMap?, organism: Organism): FileMappingPreconditionValidator {
+        submissionIdFilesMap?.values?.forEach {
+            this.validateCategoriesMatchSchema(it, organism)
+        }
+        return this
+    }
+
+    fun validateCategoriesMatchSchema(fileCategoriesFilesMap: FileCategoryFilesMap?, organism: Organism): FileMappingPreconditionValidator {
+        if (fileCategoriesFilesMap == null) return this
+        val allowedCategories = backendConfig
+            .getInstanceConfig(organism)
+            .schema.submissionDataTypes.files.categories
+            .map { it.name }
+
+        fileCategoriesFilesMap.categories().forEach { category: FileCategory ->
+            if (!allowedCategories.contains(category)) {
+                throw UnprocessableEntityException(
+                    "The category $category is not part of the configured categories for $organism.",
+                )
+            }
+        }
+        return this
+    }
+}
